improvement(falcon-server): Separating internal method to graphqlUtils
