+++ packages/falcon-server/src/schemaDirectives/GraphQLCacheDirective.js
@@ -1,16 +1,15 @@
 const { SchemaDirectiveVisitor } = require('graphql-tools');
 const { defaultFieldResolver } = require('graphql');
-const { getRootType, getOperationPath, getFieldValue, findIdFieldName, PATH_SEPARATOR } = require('../graphqlUtils');
+const { getRootType, getOperationPath, extractTagsForIdPath, getTagsForField } = require('../graphqlUtils');
 const { createShortHash } = require('../utils');
 
 // Default cache TTL (10 minutes)
 const DEFAULT_TTL = 10;
-const TAG_SEPARATOR = ':';
-const PARENT_KEYWORD = '$parent';
 
 /**
  * @typedef {import('graphql').GraphQLType} GraphQLType
  * @typedef {import('graphql').GraphQLField} GraphQLField
+ * @typedef {import('graphql').GraphQLResolveInfo} GraphQLResolveInfo
  */
 
 /**
@@ -51,7 +50,7 @@ module.exports = class GraphQLCacheDirective extends SchemaDirectiveVisitor {
   /**
    * Get a resolver function with caching capabilities (depends on the provided config)
    * @param {Function} resolve Native GQL resolver function
-   * @param {object} field Field info object
+   * @param {GraphQLField} field Field info object
    * @param {object} defaultCacheConfig Default cache config
    * @return {Function} Resolver function with caching
    */
@@ -100,7 +99,7 @@ module.exports = class GraphQLCacheDirective extends SchemaDirectiveVisitor {
    * Execute the actual GraphQL resolver and generate cache tags
    * @param {object} result Resolver result
    * @param {object} parent GraphQL parent object
-   * @param {object} info GraphQL Info object
+   * @param {GraphQLResolveInfo} info GraphQL Info object
    * @return {object} Final resolver result
    */
   handleCacheCallbackResponse(result, parent, info) {
@@ -110,10 +109,10 @@ module.exports = class GraphQLCacheDirective extends SchemaDirectiveVisitor {
     const tags = [returnTypeName];
 
     // Checking if Type is "self-cacheable"
-    tags.push(...this.getTagsForField(resolverResult, info.returnType));
+    tags.push(...getTagsForField(resolverResult, info.returnType));
 
     idPath.forEach(idPathEntry => {
-      tags.push(...this.extractTagsForIdPath(idPathEntry, resolverResult, info, parent));
+      tags.push(...extractTagsForIdPath(idPathEntry, resolverResult, info, parent));
     });
 
     return {
@@ -126,79 +125,12 @@ module.exports = class GraphQLCacheDirective extends SchemaDirectiveVisitor {
   }
 
   /**
-   * Extract cache tags for the provided ID path and return a list of tags
-   * @param {string} idPath ID operation path string (like "$parent.items" or "items")
-   * @param {object} result Resolver result
-   * @param {object} info GraphQL info object
-   * @param {object} parent GraphQL parent object
-   * @return {string[]} List of tags
-   */
-  extractTagsForIdPath(idPath, result, info, parent) {
-    const [rootPath, ...fieldPathSections] = idPath.split(PATH_SEPARATOR);
-    const valueToCheck = rootPath === PARENT_KEYWORD ? parent : result;
-    const typeToCheck = getRootType(rootPath === PARENT_KEYWORD ? info.parentType : info.returnType);
-    if (rootPath !== PARENT_KEYWORD) {
-      // Put first path section back to "fieldPathSections" for non-parent entries
-      fieldPathSections.unshift(rootPath);
-    }
-
-    return this.getTagsForField(valueToCheck, typeToCheck, fieldPathSections);
-  }
-
-  /**
-   * Get a list of tags from the provided `sourceValue` using specified `fieldType` and `fieldPathSections` (for nested values)
-   * @param {object} sourceValue Source value to get tags from
-   * @param {object} fieldType GraphQL Field Type object
-   * @param {string[]} [fieldPathSections=[]] An optional field path sections (example: ["products", "items"] which are created from a relative "products.items" field path)
-   * that are going to be used to get tags from. If not passed or empty - tags will be received from `sourceValue` directly.
-   * @return {string[]} List of tag names
-   */
-  getTagsForField(sourceValue, fieldType, fieldPathSections = []) {
-    if (!fieldPathSections.length) {
-      const { name: typeName } = getRootType(fieldType);
-      return GraphQLCacheDirective.generateTagNames(typeName, getFieldValue(sourceValue, findIdFieldName(fieldType)));
-    }
-
-    const [currentPath, ...restIdPath] = fieldPathSections;
-    const { _fields: fields } = fieldType;
-    let { name: typeName } = fieldType;
-    let fieldValue = getFieldValue(sourceValue, currentPath);
-
-    // Keep looking for nested ID path until it reaches the end node
-    if (currentPath && restIdPath.length) {
-      return this.getTagsForField(fieldValue, getRootType(fields[currentPath].type), restIdPath);
-    }
-    if (Array.isArray(fieldValue)) {
-      const currentType = getRootType(fields[currentPath].type);
-      typeName = currentType.name;
-      const currentCacheIdFieldName = findIdFieldName(currentType);
-      fieldValue = getFieldValue(fieldValue, currentCacheIdFieldName);
-    }
-
-    return [typeName, ...GraphQLCacheDirective.generateTagNames(typeName, fieldValue)];
-  }
-
-  /**
-   * Generate tag names using `entityName` and `entityId`
-   * @param {string} entityName Entity Type name (like "Product")
-   * @param {string|string[]} entityId Entity ID or list of IDs (like: "1" or ["1", "2"])
-   * @return {string[]} List of tag names (example: ["Product:1", "Product:2"])
-   */
-  static generateTagNames(entityName, entityId) {
-    if (!entityId) {
-      return [];
-    }
-    const ids = Array.isArray(entityId) ? entityId.filter(element => element) : [entityId];
-    return ids.map(id => `${entityName}${TAG_SEPARATOR}${id}`);
-  }
-
-  /**
-   * Returns cache options object based on the provided data in this order/prio:
+   * Returns cache options object based on the provided data in this order/priority:
    * - default cache config
    * - default cache config provided from `context.config`
    * - cache config provided in `@cache(...)` directive
    * - cache config for a specific operation via `context.config`
-   * @param {object} info GraphQL Request info object
+   * @param {GraphQLResolveInfo} info GraphQL Request info object
    * @param {object} resolversCacheConfig Cache object provided via `context.config`
    * @param {object} defaultDirectiveValue Default options defined in cache directive for the specific type
    * @return {object} Final cache options object
